# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST de gestion de bibliothèque universitaire.",
        "title": "Bibliothèque API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/departements": {
            "get": {
                "description": "Retourne la liste des departements d'ecole.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Lister les departements",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DepartementEcole"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/emprunts": {
            "get": {
                "description": "Retourne les emprunts actifs pour un utilisateur donne.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Lister les emprunts d'un utilisateur",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID utilisateur",
                        "name": "utilisateur_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/repositories.EmpruntItem"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Enregistre l'emprunt d'un exemplaire pour un utilisateur.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Enregistrer un emprunt",
                "parameters": [
                    {
                        "description": "Demande d'emprunt",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.EmpruntRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/emprunts/verifier": {
            "get": {
                "description": "Verifie la possibilite d'emprunt et retourne un apercu.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emprunts"
                ],
                "summary": "Verifier un emprunt",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID utilisateur",
                        "name": "utilisateur_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Code barre exemplaire",
                        "name": "code_barre",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/repositories.PreviewEmprunt"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/utilisateurs": {
            "post": {
                "description": "Cree un etudiant, un enseignant ou un particulier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Creer un utilisateur",
                "parameters": [
                    {
                        "description": "Donnees utilisateur",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CreerUtilisateurRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreerUtilisateurResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.CreerUtilisateurRequest": {
            "type": "object",
            "required": [
                "code_postal",
                "date_naissance",
                "email",
                "nom",
                "nom_rue",
                "numero_rue",
                "numero_telephone",
                "prenom",
                "statut",
                "ville"
            ],
            "properties": {
                "annee_etude": {
                    "type": "string"
                },
                "code_postal": {
                    "type": "string"
                },
                "date_naissance": {
                    "type": "string"
                },
                "departement_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "nom": {
                    "type": "string"
                },
                "nom_rue": {
                    "type": "string"
                },
                "numero_rue": {
                    "type": "string"
                },
                "numero_telephone": {
                    "type": "string"
                },
                "prenom": {
                    "type": "string"
                },
                "statut": {
                    "type": "string",
                    "enum": [
                        "etudiant",
                        "enseignant",
                        "particulier"
                    ]
                },
                "ville": {
                    "type": "string"
                }
            }
        },
        "controllers.CreerUtilisateurResponse": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "mot_de_passe": {
                    "type": "string"
                }
            }
        },
        "controllers.EmpruntRequest": {
            "type": "object",
            "required": [
                "code_barre",
                "utilisateur_id"
            ],
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "utilisateur_id": {
                    "type": "integer"
                }
            }
        },
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "erreur": {
                    "type": "string"
                }
            }
        },
...
        "controllers.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.DepartementEcole": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nom": {
                    "type": "string"
                }
            }
        },
...
        "repositories.EmpruntItem": {
            "type": "object",
            "properties": {
                "code_barre": {
                    "type": "string"
                },
                "date_debut": {
                    "type": "string"
                },
                "date_fin": {
                    "type": "string"
                },
                "en_retard": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
...
        "repositories.PreviewEmprunt": {
            "type": "object",
            "properties": {
                "caution": {
                    "type": "number"
                },
                "code_barre": {
                    "type": "string"
                },
                "nouveau_solde": {
                    "type": "number"
                },
                "solde_actuel": {
                    "type": "number"
                },
                "titre": {
                    "type": "string"
                }
            }
        },
...
    }
}